pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
]
